asprintf() is starting a coup...
git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@652 5f96b256-904b-4d8d-8c98-d829582c6739

+++ include/objects.h
@@ -180,8 +180,8 @@ typedef struct host_struct{
 	char    *perf_data;
         int     state_type;
 	int     current_attempt;
-	int     current_event_id;
-	int     last_event_id;
+	unsigned long current_event_id;
+	unsigned long last_event_id;
 	double  latency;
 	double  execution_time;
 	int     check_options;
@@ -202,7 +202,7 @@ typedef struct host_struct{
 	int     notified_on_unreachable;
 	int     current_notification_number;
 	int     no_more_notifications;
-	int     current_notification_id;
+	unsigned long current_notification_id;
 	int     check_flapping_recovery_notification;
 	int     scheduled_downtime_depth;
 	int     pending_flex_downtime;
@@ -365,8 +365,8 @@ typedef struct service_struct{
 	int     should_be_scheduled;
 	time_t	last_check;
 	int	current_attempt;
-	int     current_event_id;
-	int     last_event_id;
+	unsigned long current_event_id;
+	unsigned long last_event_id;
 	time_t	last_notification;
 	time_t  next_notification;
 	int     no_more_notifications;
@@ -383,7 +383,7 @@ typedef struct service_struct{
 	int     notified_on_warning;
 	int     notified_on_critical;
 	int     current_notification_number;
-        int     current_notification_id;
+        unsigned long current_notification_id;
 	double  latency;
 	double  execution_time;
 	int     is_executing;
